+++ R/adj_intens.R
@@ -5,8 +5,8 @@
 #'
 #' @details
 #' Many of the Open Specy functions will assume that the spectrum is in
-#' absorbance units. For example, see \code{\link{match_spectrum}()} and
-#' \code{\link{subtract_background}()}.
+#' absorbance units. For example, see \code{\link{match_spec}()} and
+#' \code{\link{subtr_bg}()}.
 #' To run those functions properly, you will need to first convert any spectra
 #' from transmittance or reflectance to absorbance using this function.
 #' The transmittance adjustment uses the \eqn{log10(1 / T)}
@@ -24,68 +24,69 @@
 #' @param type a character string specifying whether the input spectrum is
 #' in absorbance units (\code{"none"}, default) or needs additional conversion
 #' from \code{"reflectance"} or \code{"transmittance"} data.
-#' @param make_relative logical; if \code{TRUE} spectra are automatically
-#' normalized with \code{\link{make_relative}()}.
+#' @param make_rel logical; if \code{TRUE} spectra are automatically normalized
+#' with \code{\link{make_rel}()}.
 #' @param \ldots further arguments passed to the submethods.
 #'
 #' @return
-#' \code{adjust_intensity()} returns a data frame containing two columns
+#' \code{adj_intens()} returns a data frame containing two columns
 #' named \code{"wavenumber"} and \code{"intensity"}.
 #'
 #' @examples
 #' data("raman_hdpe")
-#' adjust_intensity(raman_hdpe)
+#'
+#' adj_intens(raman_hdpe)
 #'
 #' @author
 #' Win Cowger, Zacharias Steinmetz
 #'
 #' @seealso
-#' \code{\link{subtract_background}()} for spectral background correction;
-#' \code{\link{match_spectrum}()} matches spectra with the Open Specy or other
+#' \code{\link{subtr_bg}()} for spectral background correction;
+#' \code{\link{match_spec}()} matches spectra with the Open Specy or other
 #' reference libraries
 #'
 #' @importFrom magrittr %>%
 #' @export
-adjust_intensity <- function(x, ...) {
-  UseMethod("adjust_intensity")
+adj_intens <- function(x, ...) {
+  UseMethod("adj_intens")
 }
 
-#' @rdname adjust_intensity
+#' @rdname adj_intens
 #'
 #' @export
-adjust_intensity.formula <- function(formula, data = NULL, ...) {
-  if (missing(formula) || (length(formula) != 3L) || (length(attr(terms(formula[-2L]),
-                                                                  "term.labels")) != 1L))
+adj_intens.formula <- function(formula, data = NULL, ...) {
+  if (missing(formula) || (length(formula) != 3L) ||
+      (length(attr(terms(formula[-2L]), "term.labels")) != 1L))
     stop("'formula' missing or incorrect")
 
   mf <- model.frame(formula, data)
   lst <- as.list(mf)
   names(lst) <- c("y", "x")
 
-  do.call("adjust_intensity", c(lst, list(...)))
+  do.call("adj_intens", c(lst, list(...)))
 }
 
-#' @rdname adjust_intensity
+#' @rdname adj_intens
 #'
 #' @export
-adjust_intensity.data.frame <- function(x, ...) {
+adj_intens.data.frame <- function(x, ...) {
   if (!all(c("wavenumber", "intensity") %in% names(x)))
     stop("'data' must contain 2 columns named 'wavenumber' and 'intensity'")
 
-  do.call("adjust_intensity", list(x$wavenumber, x$intensity, ...))
+  do.call("adj_intens", list(x$wavenumber, x$intensity, ...))
 }
 
-#' @rdname adjust_intensity
+#' @rdname adj_intens
 #'
 #' @export
-adjust_intensity.default <- function(x, y, type = "none", make_relative = TRUE,
-                                     ...) {
+adj_intens.default <- function(x, y, type = "none", make_rel = TRUE,
+                               ...) {
   yadj <- switch(type,
-                 "reflectance" = (1 - adjust_negative(y))^2 / (2 * adjust_negative(y)),
-                 "transmittance" = log10(1/adjust_negative(y)),
-                 "none" = adjust_negative(y)
-                )
-  if (make_relative) yout <- make_relative(yadj) else yout <- yadj
+                 "reflectance" = (1 - adj_neg(y))^2 / (2 * adj_neg(y)),
+                 "transmittance" = log10(1/adj_neg(y)),
+                 "none" = adj_neg(y)
+  )
+  if (make_rel) yout <- make_rel(yadj) else yout <- yadj
 
   data.frame(wavenumber = x, intensity = yout)
 }