+++ R/data_norm.R
@@ -1,21 +1,21 @@
-#' @rdname data_normalization
+#' @rdname data_norm
 #'
 #' @title Normalization of spectral data
 #'
 #' @description
-#' \code{adjust_negative()} converts numeric values \code{x} < 1 into values
+#' \code{adj_neg()} converts numeric values \code{x} < 1 into values
 #' >= 1, keeping absolute differences between values by shifting intensity
 #' values with the value of the smallest number.
-#' \code{make_relative()} converts values \code{x} into relative values between
+#' \code{make_rel()} converts values \code{x} into relative values between
 #' 0 and 1 using the standard normalization equation.
 #' If \code{na.rm} is \code{TRUE}, missing values are removed before the
 #' computation proceeds.
 #'
 #' @details
-#' \code{adjust_negative()} is used in Open Specy to avoid errors that could
+#' \code{adj_neg()} is used in Open Specy to avoid errors that could
 #' arise from log transforming spectra when using
-#' \code{\link{adjust_intensity}()} and other functions.
-#' \code{make_relative()} is used in Open Specy to retain the relative
+#' \code{\link{adj_intens}()} and other functions.
+#' \code{make_rel()} is used in Open Specy to retain the relative
 #' height proportions between spectra while avoiding the large numbers that can
 #' result from some spectral instruments.
 #'
@@ -24,23 +24,23 @@
 #' @param na.rm logical. Should missing values be removed?
 #'
 #' @return
-#' \code{adjust_negative()} and \code{make_relative()} return numeric vectors
+#' \code{adj_neg()} and \code{make_rel()} return numeric vectors
 #' with the normalized data.
 #'
 #' @examples
-#' adjust_negative(c(-1000, -1, 0, 1, 10))
-#' make_relative(c(-1000, -1, 0, 1, 10))
+#' adj_neg(c(-1000, -1, 0, 1, 10))
+#' make_rel(c(-1000, -1, 0, 1, 10))
 #'
 #' @author
 #' Win Cowger, Zacharias Steinmetz
 #'
 #' @seealso
 #' \code{\link[base]{min}()} for the calculation of minima;
-#' \code{\link{adjust_intensity}()} for log transformation functions
+#' \code{\link{adj_intens}()} for log transformation functions
 #'
 #' @importFrom magrittr %>%
 #' @export
-adjust_negative <- function(x, na.rm = FALSE) {
+adj_neg <- function(x, na.rm = FALSE) {
   if (min(x, na.rm = na.rm) < 1) {
     x + min(x, na.rm = na.rm) %>% abs() + 1
   } else {
@@ -48,9 +48,9 @@ adjust_negative <- function(x, na.rm = FALSE) {
   }
 }
 
-#' @rdname data_normalization
+#' @rdname data_norm
 #'
 #' @export
-make_relative <- function(x, na.rm = FALSE) {
+make_rel <- function(x, na.rm = FALSE) {
   (x - min(x, na.rm = na.rm)) / (max(x, na.rm = na.rm) - min(x, na.rm = na.rm))
 }