+++ NEWS.md
@@ -4,12 +4,13 @@
 
 - Use external Open Specy libraries from OSF
 - `read_asp()` for reading Agilent .asp files
+- Comprehensive package vignette and function documentation
 - CI testing for main functions
 
 ## Minor Improvements
 
 - GUI improvements (loading screen, toggles)
-- Better error handling and documentation
+- Better error handling
 - Stripped down dependencies
 
 ## Bug fixes